Working context init

+++ src/fork/fork_pair.c
@@ -6,16 +6,16 @@
 /*   By: rorollin <rorollin@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/30 20:48:07 by rorollin          #+#    #+#             */
-/*   Updated: 2025/09/30 20:50:23 by rorollin         ###   ########.fr       */
+/*   Updated: 2025/10/01 16:18:25 by rorollin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-t_fork_pair	fork_pair_gen(t_fork *head, int index)
+t_fork_pair	fork_pair_gen(t_fork *head, size_t index)
 {
 	t_fork_pair	ret;
-	int			i;
+	size_t		i;
 
 	i = 0;
 	while (i < index)
